% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\name{bookmark_list}
\alias{bookmark_list}
\title{List all bookmark names}
\usage{
bookmark_list()
}
\value{
A character vector of bookmark names, or a message if none found.
}
\description{
List all bookmark names
}
