% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\name{bookmark_read}
\alias{bookmark_read}
\title{Read the content of a specific bookmark by name}
\usage{
bookmark_read(name)
}
\arguments{
\item{name}{The name of the bookmark to read.}
}
\value{
The content of the bookmark, or a message if not found.
}
\description{
Read the content of a specific bookmark by name
}
