% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\name{bookmark_save}
\alias{bookmark_save}
\title{Save or update a named bookmark with some text content}
\usage{
bookmark_save(name, content)
}
\arguments{
\item{name}{The name of the bookmark.}

\item{content}{The text to save in the bookmark.}
}
\value{
A message indicating the bookmark was saved.
}
\description{
Save or update a named bookmark with some text content
}
