% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckduckgo-instant-answer.R
\name{duckduckgo_answer}
\alias{duckduckgo_answer}
\title{Call DuckDuckGo Instant Answer API}
\usage{
duckduckgo_answer(
  query,
  no_redirect = FALSE,
  no_html = FALSE,
  skip_disambig = FALSE,
  app_name = "duckduckr"
)
}
\arguments{
\item{query}{the query string}

\item{no_redirect}{TRUE to skip HTTP redirects (for !bang commands)}

\item{no_html}{TRUE to remove html from results}

\item{skip_disambig}{TRUE to skip disambiguation (D) Type.}

\item{app_name}{the appname used to identify your application.}
}
\value{
A list. Includes parsed response or error details as attributes.
}
\description{
Makes a synchronous API call to the DuckDuckGo Instant Answer API.
}
