% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_time.R
\name{get_current_time}
\alias{get_current_time}
\title{Gets the current time in the given time zone.}
\usage{
get_current_time(tz = "UTC")
}
\arguments{
\item{tz}{The time zone to get the current time in.}
}
\value{
The current time in the given time zone.
}
\description{
Gets the current time in the given time zone.
}
