% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_wd.R
\name{get_current_wd}
\alias{get_current_wd}
\title{Get the current working directory as a JSON object}
\usage{
get_current_wd()
}
\value{
A JSON string representing the files in the current working directory.
}
\description{
Get the current working directory as a JSON object
}
