% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_current_forecast}
\alias{get_current_forecast}
\title{Get NWS Forecast from Latitude and Longitude}
\usage{
get_current_forecast(
  lat,
  lon,
  forecast_type = "general",
  forecast_detail = "short"
)
}
\arguments{
\item{lat}{Numeric. Latitude of the location (e.g., 40.886).}

\item{lon}{Numeric. Longitude of the location (e.g., -81.416).}

\item{forecast_type}{Character. Forecast frequency granularity which can be either `"daily"` (default) or `"hourly"`.}

\item{forecast_detail}{Character. Forecast level of detail which can be a concise summary or a full description. It can be either `"short"` (default) or `"detailed"`.}
}
\value{
A list containing the forecast data as returned by the NWS API.
}
\description{
Retrieves weather forecast data (either general or hourly) from the
National Weather Service (NWS) API for a given latitude and longitude.
}
\examples{
\dontrun{
get_current_forecast(40.8860529, -81.4164969, "general", "short")
get_current_forecast(40.8860529, -81.4164969, "hourly", "detailed")
}
}
