% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckduckgo.R
\name{\%||\%}
\alias{\%||\%}
\title{Internal utility: Provide a default value if the first argument is NULL}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value to check for NULL}

\item{b}{Value to return if `a` is NULL}
}
\value{
`a` if not NULL, otherwise `b`
}
\description{
Internal utility: Provide a default value if the first argument is NULL
}
\examples{
\dontrun{
NULL \%||\% "default"
5 \%||\% "default"
}
}
\keyword{internal}
