% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia.R
\name{html_string_as_markdown}
\alias{html_string_as_markdown}
\title{Internal: Convert HTML string to Markdown}
\usage{
html_string_as_markdown(x)
}
\arguments{
\item{x}{HTML string}
}
\value{
Markdown string
}
\description{
Internal: Convert HTML string to Markdown
}
\keyword{internal}
