% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarks.R
\docType{data}
\name{tool_bookmark_read}
\alias{tool_bookmark_read}
\title{Tool to help register the `bookmark_read` function with ellmer}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_bookmark_read
}
\description{
Tool to help register the `bookmark_read` function with ellmer
}
\keyword{datasets}
