% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_mini_chat.R
\docType{data}
\name{tool_call_mini_chat}
\alias{tool_call_mini_chat}
\title{Tool to Call GPT-4o Mini Chat via ellmer}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_call_mini_chat
}
\description{
Registers `call_mini_chat()` as an ellmer tool to interact with the GPT-4o mini model.
}
\keyword{datasets}
