% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckduckgo-instant-answer.R
\docType{data}
\name{tool_duckduckgo_answer}
\alias{tool_duckduckgo_answer}
\title{Tool to call DuckDuckGo Instant Answer API}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_duckduckgo_answer
}
\description{
Tool to call DuckDuckGo Instant Answer API
}
\keyword{datasets}
