% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckduckgo.R
\docType{data}
\name{tool_duckduckgo_news}
\alias{tool_duckduckgo_news}
\title{Tool to search DuckDuckGo for news articles and return results}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_duckduckgo_news
}
\description{
Tool to search DuckDuckGo for news articles and return results
}
\keyword{datasets}
