% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_r_code_chunks.R
\docType{data}
\name{tool_extract_r_code_chunks}
\alias{tool_extract_r_code_chunks}
\title{Tool to extract R code chunks from Markdown using ellmer}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_extract_r_code_chunks
}
\description{
Tool to extract R code chunks from Markdown using ellmer
}
\keyword{datasets}
