% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_weather.R
\docType{data}
\name{tool_get_current_forecast}
\alias{tool_get_current_forecast}
\title{Tools to help register the `get_current_forecast` function with ellmer}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_get_current_forecast
}
\description{
Tools to help register the `get_current_forecast` function with ellmer
}
\keyword{datasets}
