% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_wd.R
\docType{data}
\name{tool_get_current_wd}
\alias{tool_get_current_wd}
\title{Tool to help register the `get_current_wd` function with ellmer}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_get_current_wd
}
\description{
Tool to help register the `get_current_wd` function with ellmer
}
\keyword{datasets}
