% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\docType{data}
\name{tool_read_clipboard}
\alias{tool_read_clipboard}
\title{Tool to read and write clipboard content}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_read_clipboard
}
\description{
Tool to read and write clipboard content
}
\keyword{datasets}
