% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_content.R
\docType{data}
\name{tool_retrieve_url_content}
\alias{tool_retrieve_url_content}
\title{Tool to read URL content as Markdown}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_retrieve_url_content
}
\description{
Tool to read URL content as Markdown
}
\keyword{datasets}
