% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_r_code.R
\docType{data}
\name{tool_run_r_code}
\alias{tool_run_r_code}
\title{Tool to run R code provided as a string and return the result}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_run_r_code
}
\description{
Tool to run R code provided as a string and return the result
}
\keyword{datasets}
