% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia.R
\docType{data}
\name{tool_wikipedia_get_page_summary}
\alias{tool_wikipedia_get_page_summary}
\title{Tool to get a Wikipedia page summary}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_wikipedia_get_page_summary
}
\description{
Tool to get a Wikipedia page summary
}
\keyword{datasets}
