% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\docType{data}
\name{tool_write_to_clipboard}
\alias{tool_write_to_clipboard}
\title{Tool to help register the `tool_read_clipboard` function with ellmer}
\format{
An object of class \code{ellmer::ToolDef} (inherits from \code{S7_object}) of length 1.
}
\usage{
tool_write_to_clipboard
}
\description{
Tool to help register the `tool_read_clipboard` function with ellmer
}
\keyword{datasets}
